Extract the A/B-test-wrapped toggle into a component
